backend/requirements/base.txt
Django==5.1.6
djangorestframework==3.15.2
django-environ==0.11.2
django-cors-headers==4.6.0
channels==4.1.0
channels-redis==4.2.0
requests==2.32.3
pydantic==2.9.2
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
daphne==4.1.2